install GPGMail plugin outside Mail bundle folder
The directory ~/Library/Mail/Bundles is protected by sandboxing and the
XPC service will not work in there. But placing components outside does
not help either, because the plugin needs to access things from within
the Mail sandbox.

The official GPGTools installation solves this by installing components
in system-wide locations that can be read from within any sandbox. My
solution is to install everything in ~/Library/Application Support/GPGTools
and a copy in ~/Library/Mail/Bundles.