install directy in the Mail bundle folder again
The framework can go to ~/Library/Components, which is accessible from
within the Mail app sandbox and also from the outside by the XPC
service.

Use PlistBuddy instead of defaults to modify plists in place.

Also, explicit rpath-linkage to the bundle framework directory is not
necessary, because this directory is searched implicitly.